requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
google-generativeai
openpyxl
pandas
python-multipart
python-dotenv
alembic
